requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
cffi==1.14.1
chardet==3.0.4
discord.py==1.4.1
idna==2.10
multidict==4.7.6
pkg-resources==0.0.0
pycparser==2.20
PyNaCl==1.4.0
python-dotenv==0.14.0
six==1.15.0
typing-extensions==3.7.4.2
yarl==1.5.1
youtube-dl==2020.7.28
